Add unit tests for use case in charge of creating merchants
In the same fashion than before with the acceptance test, I'm following
the approach suggested by Codely, where the use case is considered to be
the unit and the infrastructure (repository, I/O) is mocked/stubbed.

Normally I'd use a batch to enqueue all related jobs, but it's a
functionality only available with Sidekiq PRO. There are alternative
gems that don't work quite the same.

These tests are currently in green.

References:

* https://github.com/sidekiq/sidekiq/wiki/Batches
* https://github.com/breamware/sidekiq-batch
* https://docs.gitlab.com/ee/development/sidekiq/logging.html
* https://dry-rb.org/gems/dry-types/1.7/built-in-types/
* https://github.com/rails/rails/blob/main/activesupport/lib/active_support/core_ext/numeric/bytes.rb#L5
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecreturnfromstub