Add triggers to update orders when disbursements are created or updated
A disbursement probably should not be modified after it's created, but
anyway, the code takes that possibility into account.

With this approach we are coupled to the database to update orders'
disbursement ID when a disbursement is created or updated. Ideally we
could publish a domain event that eventually would update expected
records (eventual consistency).

References:

* https://www.postgresql.org/docs/current/sql-createfunction.html
* https://www.postgresql.org/docs/current/sql-createtrigger.html
* https://sequel.jeremyevans.net/rdoc-plugins/files/lib/sequel/extensions/pg_array_rb.html
* https://github.com/jeremyevans/sequel/blob/master/doc/querying.rdoc
* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsblocklength