Allow to use a fake in memory bus in tests
This change is specially important because in acceptance tests the code
would continue endlessly until no more use cases would publish a new
event.

Example:

When an order is created, an event is published. An event subscriber is
listening to that event so when it is received, an order commission is
created. In the same way, when that order commission is created, another
event will be published. Assumming another event subscriber is listening
to that other event, the process wouldn't finish until no event is published
or a published event has no subscribers.

I'm using RSpec hooks to enable the fake in memory event bus in certain
tests. For that purpose, add the label :fake_event_bus to an example.

References:

* https://github.com/rspec/rspec-core/blob/main/lib/rspec/core/hooks.rb
* https://dry-rb.org/gems/dry-system/1.0/test-mode/