Add async event bus
This new bus enqueues a generic job with Sidekiq.

Two arguments must be provided:

* The name of the event subscriber (FQN, Fully Qualified Name).
* The serialized event in JSON format.

A dedicated Sidekiq queue has been created as well (domain_events).

In addition, adds more flexibility in tests, so a specific event bus can
be faked, and use cases, where a specific event bus can be injected.

A Ruby JSON schema validator is used to validate the format of the event
that is serialized (via gem comforming to JSON Schema Draft 6). In case
is not valid, an exception is raised with the errors found.

References:

* https://github.com/voxpupuli/json-schema
* https://github.com/voxpupuli/json-schema/blob/v4.1.1/lib/json-schema/validator.rb
* https://github.com/sidekiq/sidekiq/wiki/Advanced-Options#the-sidekiq-configuration-file
* https://en.wikipedia.org/wiki/Fully_qualified_name#Examples