Allow to create a commission when an order is created
The event about an order created is published from the event bus to the
event subscriber so the commission can be created.

The commission is created as long as following conditions are met:

* The order exists.
* The order has not have any commission associated yet.

If any of them is not met, then a domain exception is raised.

Two services has been added with following responsibilities:

* Extract applicable fee based on the order amount.
* Calculate the commission amount based on the fee and the order amount.

The event subscriber tests are now in green.

Unit tests for the use case in charge of creating an order commission
are in green as well.

The only tests missing are the integration ones for the repository.

References:

* https://rom-rb.org/learn/sql/3.3/queries/
* https://api.rom-rb.org/rom-sql/ROM/SQL/Relation/Reading#exist%3F-instance_method
* https://stackoverflow.com/questions/17534230/is-it-possible-to-use-a-range-as-a-key-for-a-hash-in-ruby
* https://ruby-doc.org/3.2.2/Range.html#class-Range-label-Endless+Ranges
* https://stackoverflow.com/questions/59850951/when-to-use-float-vs-decimal
* https://docs.rubocop.org/rubocop/cops_style.html#stylecaseequality