Allow acceptance test for generating disbursements to take monthly fees into account
Monthly fees will be generated only after first disbursement of the
month is generated for a merchant, as long as the minimum monthly fee
has not been covered with order commissions.

That monthly fee will be calculated and stored but not substracted from
the amount disbursed to the merchant. For that reason, it is currently
calculated after the disbursement has been created.

If those requirements change, that part of the code will have to change.

This test acts as a kind of e2e test that ensures the entire process of
generating disbursements and monthly fees works as expected.

It is currently in red.